Add gcd algorithm
We will use gcd for reducing Rationals to simplest form on
construction.

gcd makes use of std::enable_if to detect if the input type
is integral. This way, gcd will handle int, long, and
long long in the same way. This is necessary because c++'s
modulus may return a negative number...which is not what
we want.